Consolidate regexp
